fs/archive: move file and directory classes to their own files
